+++ flutter_searchbox_ui/example/lib/main.dart
@@ -54,7 +54,7 @@ class HomePage extends StatelessWidget {
         primarySwatch: Colors.blue,
         visualDensity: VisualDensity.adaptivePlatformDensity,
       ),
-       home: Scaffold(
+      home: Scaffold(
           appBar: AppBar(
             title: Text('RangeInput Demo'),
           ),
@@ -89,31 +89,87 @@ class HomePage extends StatelessWidget {
                 child: Center(
                   child: RangeInput(
                     id: 'range-selector',
-                    title: "Publication Year",
-                    rangeLabel: "to",
+                    buildTitle: () {
+                      return const Text(
+                        "Custom Title Widget",
+                        style: TextStyle(
+                          fontWeight: FontWeight.bold,
+                          fontSize: 16.0,
+                          color: Colors.amber,
+                        ),
+                      );
+                    },
+                    buildRangeLabel: () {
+                      return const Text(
+                        "unless",
+                        style: TextStyle(
+                          fontWeight: FontWeight.bold,
+                          fontSize: 16.0,
+                          color: Colors.blue,
+                        ),
+                      );
+                    },
                     dataField: 'original_publication_year',
-                    range: RangeType(
-                      start: 3000,
-                      end: ['other', 1990, 2000, 2010],
+                    range: const RangeType(
+                      start: 1900,
+                      end: ['other', 1990, 2000, 2010, 'no limit'],
                     ),
-                    defaultValue: DefaultValue(start: 1980, end: 2060),
+                    defaultValue: const DefaultValue(start: 1980, end: 2000),
                     rangeLabels: RangeLabelsType(
                       start: (value) {
-                        return value == 'other' ? 'Custom Other' : 'yr $value';
+                        return value == 'other'
+                            ? 'Custom Other'
+                            : (value == 'no limit'
+                                ? 'No Limits custom'
+                                : 'yr $value');
                       },
                       end: (value) {
-                        return value == 'other' ? 'Custom Other' : 'yr $value';
+                        return value == 'other'
+                            ? 'Custom Other'
+                            : (value == 'no limit'
+                                ? 'No Limits custom'
+                                : 'yr $value');
                       },
                     ),
                     validateRange: (start, end) {
-                      print('custom validate');
                       if (start < end) {
                         return true;
                       }
                       return false;
                     },
-                    errorMessage: (start, end) {
-                      return 'Custom error $start > $end';
+                    buildErrorMessage: (start, end) {
+                      return Text(
+                        'Custom error $start > $end',
+                        style: const TextStyle(
+                          fontSize: 15.0,
+                          color: Colors.yellowAccent,
+                        ),
+                      );
+                    },
+                    inputStyle: const TextStyle(
+                      fontSize: 18,
+                      height: 1,
+                      color: Colors.deepPurple,
+                    ),
+                    dropdownStyle: const TextStyle(
+                      fontSize: 18,
+                      height: 1,
+                      color: Colors.deepPurpleAccent,
+                    ),
+                    customContainer: (showError, childWidget) {
+                      return Container(
+                        padding: const EdgeInsets.all(8.0),
+                        height: 50,
+                        decoration: BoxDecoration(
+                          border: Border.all(
+                            color:
+                                showError ? Colors.orangeAccent : Colors.black,
+                            width: 1.5,
+                          ),
+                          borderRadius: BorderRadius.circular(3),
+                        ),
+                        child: childWidget,
+                      );
                     },
                   ),
                 ),